Begin moving generic portions of peripheral, measurement, and ble state machine to separate folder
